Upgrade to latest parent pom and Jenkins with Java 8
Jenkins 2.60.1 is the first version that requires Java 8.
Jenkins 2.60.3 is the current version that is required by JCasC.

The parent pom replaced FindBugs with SpotBugs, therefore changing the
respective property.

The dependencies 'jenkins-war', 'jenkins-core' and 'jenkins-test-harness'
don't have to be specified explicitly. They are already defined in the
parent pom.

The plugins 'animal-sniffer-maven-plugin' and 'java18' also don't have
to be specified explicitly. The parent pom takes care of this.

The 'jacoco' plugin dependency version has to be increased slightly.
Otherwise the autogenerated InjectedTest fails, reporting that the
'maven-plugin' is missing.

The <dependencyManagement> section is used to enforce upper bounds
for dependencies reported by the maven-enforcer-plugin.