Refactor InfluxDB Clients and Points including integration tests (#189)
* refactor: Wrap the InfluxDB client in a re-usable wrapper class

* refactor: Wrap influxdb points in `AbstractPoint` class

* feat: Add jenkins proxy support for v2.X servers

* fix: Make InfluxDB v1.X version check more robust

The InfluxDB v1.X `/ping` API response [returns the InfluxDB version as `1.X` instead of `v1.X`](https://docs.influxdata.com/influxdb/v1/tools/api/#example)
Unless, of course, you are using InfluxDB v1.11 where the API returns `v1.11.X`...

* build(test): Add `testcontainers` and related packages as test dependencies

* build!: Make `metrics` plugin a non-optional dependency

It is not possible to test against an optional plugin dependency. See
https://www.jenkins.io/doc/developer/testing/#working-with-pipeline for
more details. The only workaround would be to use the `@WithPlugin`
decorator, but this would require the plugin file to be downloaded in the
test and I think this is kind of hacky...

* test: Add integration test